requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles>=23.1.0
httpx>=0.24.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-generativeai==0.8.5
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.1
requests>=2.31.0
pillow>=10.0.0
google-cloud-storage>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4